requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
azure-identity>=1.19.0
azure-storage-blob>=12.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1